tests: initial api tests
