+++ json_merger/merger.py
@@ -22,31 +22,10 @@
 # waive the privileges and immunities granted to it by virtue of its status
 # as an Intergovernmental Organization or submit itself to any jurisdiction.
 
-
-"""Module tests."""
+"""Invenio module that is able to merge json record objects."""
 
 from __future__ import absolute_import, print_function
 
-from flask import Flask
-from flask_babelex import Babel
 
-from json_merger import JsonMerger
-
-
-def test_version():
-    """Test version import."""
-    from json_merger import __version__
-    assert __version__
-
-
-def test_init():
-    """Test extension initialization."""
-    app = Flask('testapp')
-    ext = JsonMerger(app)
-    assert 'json-merger' in app.extensions
-
-    app = Flask('testapp')
-    ext = JsonMerger()
-    assert 'json-merger' not in app.extensions
-    ext.init_app(app)
-    assert 'json-merger' in app.extensions
+def merge_records(src, update):
+    return src
