+++ tests/conftest.py
@@ -30,6 +30,8 @@ from __future__ import absolute_import, print_function
 import pytest
 from flask import Flask
 
+from json_merger import JsonMerger
+
 
 @pytest.fixture()
 def app():
@@ -38,4 +40,5 @@ def app():
     app.config.update(
         TESTING=True
     )
+    JsonMerger(app)
     return app
